fix a batch of clang-tidy checks
misc-unused-parameters, bugprone-multi-level-implicit-pointer-conversion,
modernize-return-braced-init-list, readability-math-missing-parentheses,
bugprone-implicit-widening-of-multiplication-result and
bugprone-narrowing-conversions are now clean in our code (remaining hits
are in ext/ and the WebView2 SDK headers).

prompt: fix clang-tidy misc-unused-parameters / bugprone-multi-level-implicit-pointer-conversion / modernize-return-braced-init-list / readability-math-missing-parentheses / bugprone-implicit-widening-of-multiplication-result / bugprone-narrowing-conversions

+++ src/MobiDoc.cpp
@@ -865,7 +865,7 @@ bool MobiDoc::LoadForPdbReader(PdbReader* pdbReader) {
 
     ReportIf(len(doc) != 0);
     doc.Reset();
-    doc.cap = (u32)docUncompressedSize; // capacity hint, same trick as ByteWriter ctor
+    doc.cap = docUncompressedSize; // capacity hint, same trick as ByteWriter ctor
     int nFailed = 0;
     for (int i = 1; i <= docRecCount; i++) {
         if (!LoadDocRecordIntoBuffer(i, doc)) {
